+++ images/linux/scripts/installers/nginx.sh
@@ -7,8 +7,10 @@
 # Install Nginx
 apt-get install nginx -y
 
-# Disable nginx.service
-systemctl is-active --quiet nginx.service && systemctl stop nginx.service
-systemctl disable nginx.service
+if [[ ! -f /run/systemd/container ]]; then
+    # Disable nginx.service
+    systemctl is-active --quiet nginx.service && systemctl stop nginx.service
+    systemctl disable nginx.service
+fi
 
 invoke_tests "WebServers" "Nginx"
